+++ FerrumCore/FeCore/Strings/StringSlice.cpp
@@ -27,7 +27,7 @@ namespace FE
             }
             else
             {
-                return { ParseErrorCode::InvalidSyntax, static_cast<size_t>(ptr - Data()) };
+                return { ParseErrorCode::InvalidSyntax, static_cast<uint32_t>(ptr - Data()) };
             }
 
             ++ptr;
@@ -76,7 +76,7 @@ namespace FE
             }
             else
             {
-                return { ParseErrorCode::InvalidSyntax, static_cast<size_t>(ptr - Data()) };
+                return { ParseErrorCode::InvalidSyntax, static_cast<uint32_t>(ptr - Data()) };
             }
 
             ++ptr;
@@ -98,7 +98,7 @@ namespace FE
             slice = StringSlice(slice.Data() + 1, slice.Size() - 1);
         }
 
-        size_t position = 0;
+        uint32_t position = 0;
         for (auto c : slice)
         {
             if (c == '.')
